Move the configuration code to a separate file
This code will be reused by the webhook receiver in a future commit.

This also improves the test coverage.
