+++ src/syncer.test.ts
@@ -66,104 +66,6 @@ afterEach(() => {
   dirObj.removeCallback();
 });
 
-test("Syncer.constructor", () => {
-  expect(() => new Syncer(config, "some private key")).not.toThrowError();
-});
-
-test("Syncer.constructor invalid appID", () => {
-  config.appID = null;
-  expect(() => new Syncer(config, "some private key")).toThrowError(
-    'The configuration\'s "appID" must be set as a number',
-  );
-});
-
-test("Syncer.constructor invalid logLevel", () => {
-  config.logLevel = 123;
-  expect(() => new Syncer(config, "some private key")).toThrowError(
-    'The configuration\'s optional "logLevel" must be a string',
-  );
-});
-
-test("Syncer.constructor invalid privateKeyPath type", () => {
-  config.privateKeyPath = 123;
-  expect(() => new Syncer(config)).toThrowError('The configuration\'s "privateKeyPath" must be a string');
-});
-
-test("Syncer.constructor invalid privateKeyPath path", () => {
-  config.privateKeyPath = "/this/does/not/exist";
-  expect(() => new Syncer(config)).toThrowError(
-    'The configuration\'s "privateKeyPath" must be a valid path to a file that exists',
-  );
-});
-
-test("Syncer.constructor upstreamMappings not set", () => {
-  config.upstreamMappings = undefined;
-  expect(() => new Syncer(config, "some private key")).toThrowError(
-    'The configuration\'s "upstreamMappings" must be a valid object',
-  );
-});
-
-test("Syncer.constructor upstreamMappings wrong type", () => {
-  config.upstreamMappings = "some mapping";
-  expect(() => new Syncer(config, "some private key")).toThrowError(
-    'The configuration\'s "upstreamMappings" must be a valid object',
-  );
-});
-
-test("Syncer.constructor upstreamMappings wrong type", () => {
-  config.upstreamMappings = "some mapping";
-  expect(() => new Syncer(config, "some private key")).toThrowError(
-    'The configuration\'s "upstreamMappings" must be a valid object',
-  );
-});
-
-test("Syncer.constructor upstreamMappings.targetOrg wrong type", () => {
-  config.upstreamMappings.targetOrg = 123;
-  expect(() => new Syncer(config, "some private key")).toThrowError(
-    'The configuration\'s upstreamMappings["targetOrg"] must be a valid object',
-  );
-});
-
-test("Syncer.constructor upstreamMappings.targetOrg.upstreamOrg wrong type", () => {
-  config.upstreamMappings.targetOrg = { upstreamOrg: 123 };
-  expect(() => new Syncer(config, "some private key")).toThrowError(
-    'The configuration\'s upstreamMappings["targetOrg"]["upstreamOrg"] must be a valid object',
-  );
-});
-
-test("Syncer.constructor upstreamMappings.targetOrg.upstreamOrg.branchMappings wrong type", () => {
-  config.upstreamMappings.targetOrg = { upstreamOrg: { branchMappings: 123 } };
-  expect(() => new Syncer(config, "some private key")).toThrowError(
-    'The configuration\'s upstreamMappings["targetOrg"]["upstreamOrg"]["branchMappings"] must be a valid object',
-  );
-});
-
-test("Syncer.constructor upstreamMappings.targetOrg.upstreamOrg.branchMappings.upstreamBranch empty string", () => {
-  config.upstreamMappings.targetOrg = { upstreamOrg: { branchMappings: { upstreamBranch: "" } } };
-  expect(() => new Syncer(config, "some private key")).toThrowError(
-    'The configuration\'s upstreamMappings["targetOrg"]["upstreamOrg"]["branchMappings"]["upstreamBranch"] ' +
-      "must be a non-empty string",
-  );
-});
-
-test("Syncer.constructor upstreamMappings.targetOrg.upstreamOrg.branchMappings.upstreamBranch wrong type", () => {
-  config.upstreamMappings.targetOrg = { upstreamOrg: { branchMappings: { upstreamBranch: 123 } } };
-  expect(() => new Syncer(config, "some private key")).toThrowError(
-    'The configuration\'s upstreamMappings["targetOrg"]["upstreamOrg"]["branchMappings"]["upstreamBranch"] ' +
-      "must be a non-empty string",
-  );
-});
-
-test("Syncer.constructor upstreamMappings.targetOrg.upstreamOrg.branchMappings duplicate target branch", () => {
-  config.upstreamMappings.targetOrg = {
-    upstreamOrg: { branchMappings: { upstreamBranch: "main", otherUpstreamBranch: "main" } },
-  };
-  expect(() => new Syncer(config, "some private key")).toThrowError(
-    'The configuration\'s upstreamMappings["targetOrg"]["upstreamOrg"]["branchMappings"] contains duplicate ' +
-      "target branches",
-  );
-});
-
 test("Syncer.closePR", async () => {
   const mockClient = jest.fn();
   mockClient.pulls = jest.fn();
